refactor scraper functions
